new variance method suprisingly close to old variance method

+++ src/cdot.py
@@ -129,7 +129,7 @@ def pim_dyn(x, w, y_shape, lut_var, lut_rpr, alloc, params):
     x = np.ascontiguousarray(x, np.int32)
     w = np.ascontiguousarray(w, np.int32)
     y = np.ascontiguousarray(y, np.int32)
-    lut_var = np.ascontiguousarray(lut_var, np.int32)
+    lut_var = np.ascontiguousarray(lut_var, np.float32)
     lut_rpr = np.ascontiguousarray(lut_rpr, np.int32)
     metrics = np.ascontiguousarray(metrics, np.int64)
